setted up page redirecting and order price in decimal format

+++ admin/deleteChef.php
@@ -1,6 +1,6 @@
 <?php
 include "db-connection.php";
-if(isset($_POST['chef_id']))
+if(isset($_POST['chef_id']) && $_SESSION['role'] =='-1')
 {
     $chef_id = mysqli_real_escape_string($conn, $_POST['chef_id']);
     $query = "DELETE FROM chefs WHERE id = '$chef_id'";
@@ -8,5 +8,11 @@ if(isset($_POST['chef_id']))
         header("Location: chefDetails.php"); // Redirect after success
         exit();
     }
+}
+else{
+    
+    header("Location: ../404.php");
+    exit();
+
 }
 ?>
